+++ docs/fr_FR/changelog.md
@@ -12,6 +12,7 @@
 -   Première version stable
 
 ### 12/06/2019
+
 -   Changement des commandes pour la position des animaux, elles sont maintenant binaires
 avec 0 = extérieur et 1 = intérieur.
 -   Nouveau widget pour les animaux.
@@ -19,3 +20,14 @@ avec 0 = extérieur et 1 = intérieur.
 dans le répertoire /data/.
 -   Après l'exécution d'une commande action, la commande info correspondante est mise à jour et le
 widget est actualisé.
+
+### 13/08/2019
+
+- Compatibilité php 7.3 et font awesome 5
+- Version minimum de Jeedom requise : 3.3
+
+### 26/08/2019
+
+- Commandes action pour autoriser ou interdire de sortie un animal sur une chatière
+- Lien avec le plugin Agenda
+- Nouvelle commande info donnant la date et l'heure de la dernière entrée/sortie d'un animal